+++ src/Crm/CrmBaseHelper.fs
@@ -140,10 +140,37 @@ let getAllOptionSetMetadata proxy =
   let resp = getResponse<RetrieveAllOptionSetsResponse> proxy request
   resp.OptionSetMetadata
 
-// Retrieve specific entity metadata
+// Find relationship intersect entities
+let findRelationEntities allLogicalNames (metadata:EntityMetadata[]) =
+  metadata
+  |> Array.Parallel.map (fun md ->
+    md.ManyToManyRelationships
+    |> Array.filter (fun m2m ->
+      not(Set.contains m2m.IntersectEntityName allLogicalNames))
+    |> Array.map (fun m2m -> m2m.IntersectEntityName))
+  |> Array.concat
+  |> Array.distinct
+
+// Retrieve specific entity metadata along with any intersect
 let getSpecificEntitiesAndDependentMetadata proxy logicalNames =
   // TODO: either figure out the best degree of parallelism through code, or add it as a setting
-  getEntityMetadataBulk proxy logicalNames
+  let entities = getEntityMetadataBulk proxy logicalNames
+
+  let set = logicalNames |> Set.ofArray
+  let needActivityParty =
+    not (set.Contains "activityparty") &&
+    entities 
+    |> Array.exists (fun m -> 
+      m.Attributes 
+      |> Array.exists (fun a -> 
+        a.AttributeTypeName = AttributeTypeDisplayName.PartyListType))
+
+  let additionalEntities = 
+    findRelationEntities set entities
+    |> if needActivityParty then Array.append [|"activityparty"|] else id
+    |> getEntityMetadataBulk proxy
+
+  Array.append entities additionalEntities
   |> Array.distinctBy (fun e -> e.LogicalName)
   |> Array.sortBy (fun e -> e.LogicalName)